requirements.txt
pymongo
flask
python-dotenv
streamlit
youtube-transcript-api
google-generativeai
googletrans
pytube
google-cloud-speech
requests
git+https://github.com/openai/whisper.git  
torch>=2.0.0
torchvision
torchaudio
python-docx
pandas
fpdf
dnspython
reportlab
flask_cors
celery==5.2.7
redis==4.1.0
openpyxl
xlrd
kombu==5.2.3
bs4
beautifulsoup4
PyPDF2
Werkzeug
Gunicorn
wikipedia
pytube